# laravelProject/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión de Vehículos",
        "description": "API RESTful para gestionar marcas de vehículos, personas y su relación de propiedad con vehículos. Incluye operaciones CRUD completas y endpoints de relaciones.",
        "contact": {
            "email": "dumar.pabon@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de la API"
        }
    ],
    "paths": {
        "/api/marcas": {
            "get": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Lista todas las marcas",
                "description": "Lista todas las marcas.",
                "operationId": "cdf1f15837825105957dc1b6581c9377",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "description": "Número de registros a saltar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número máximo de registros a retornar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de marcas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Marca"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Crea una nueva marca",
                "description": "Crea una nueva marca.",
                "operationId": "011cfb0f49c0cd34dc812d3c5e40e41f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MarcaCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Marca creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Marca"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/marcas/{id}": {
            "get": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Obtiene una marca por ID",
                "description": "Muestra una marca específica.",
                "operationId": "f044df503607a332772744171e7aebcc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Marca"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Actualiza una marca",
                "description": "Actualiza una marca.",
                "operationId": "34d9e20c71578f8affd3a8a0c877d47c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MarcaUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marca actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Marca"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Elimina una marca",
                "description": "Elimina una marca.",
                "operationId": "a11c924f91478a8599debeaa76b10428",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Marca eliminada exitosamente"
                    },
                    "404": {
                        "description": "Marca no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Marca": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre_marca": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Japón"
                    }
                },
                "type": "object"
            },
            "MarcaCreate": {
                "required": [
                    "nombre_marca",
                    "pais"
                ],
                "properties": {
                    "nombre_marca": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Japón"
                    }
                },
                "type": "object"
            },
            "MarcaUpdate": {
                "properties": {
                    "nombre_marca": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Japón"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}